BUG: Regression: "E121: Undefined variable: l:idx" on ]q of last item (after bdc6bad)
If the for loop isn't entered at all, the loop variable isn't initialized.
For the initialization before the loop, and use that variable as the start value. What we really need here is a for(;;)-style loop, but Vim doesn't have this. The alternative would be a while loop, but it's easy there to forget the incrementing and be stuck in an endless loop.

(This corrects commit bdc6bad0d849fade894d8dbb1098b53798fe9053)